feat: read/write initial data
