+++ .eslintrc.json
@@ -17,5 +17,5 @@
       0
     ]
     
-    }
+  }
 }
