+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "out-of-the-maze",
+  "name": "maze-cli",
   "version": "1.0.0",
   "description": "ITRex Group Node.js Lab | Test task #2",
-  "main": "src/index.js",
+  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -14,7 +14,7 @@
   "license": "ISC",
   "preferGlobal": true,
   "bin": {
-    "maze-cli": "src/index.js"
+    "maze-cli": "index.js"
   },
   "bugs": {
     "url": "https://github.com/kastrubait/Go-through-the-maze/issues"
