+++ src/input.js
@@ -1,0 +1,12 @@
+const MAZE = [
+  ['#','#','#','#','#','#','#','#','#'],
+  ['#','+','+','+','#','+','+','+','#'],     
+  ['#','+','#','+','#','+','#','+','#'],    
+  ['+','+','#','+','0','+','#','+','#'],     
+  ['#','#','#','+','#','#','#','#','#'],     
+  ['#','#','+','+','#','#','#','#','#'],    
+  ['#','#','+','#','#','#','#','#','#'],     
+  ['#','#','#','#','#','#','#','#','#'],
+];
+
+module.exports = { MAZE };
